`validate` and `apply` cli commands prepare Settings
