+++ core/src/modules/plugin.ts
@@ -1,5 +1,4 @@
 import { api } from "actionhero";
-import { Op } from "sequelize";
 import { GrouparooPlugin } from "../classes/plugin";
 import { MustacheUtils } from "./mustacheUtils";
 import { APMWrap } from "./apm";
@@ -76,14 +75,6 @@ export namespace plugin {
   export function mountModels() {
     models.map((model) => {
       if (!model.isInitialized) api.sequelize.addModels([model]);
-    });
-  }
-
-  export async function cleanSettings(inUseKeys: string[]) {
-    return await Setting.destroy({
-      where: {
-        key: { [Op.notIn]: inUseKeys },
-      },
     });
   }
